Make sure cursor is trailing within opaque inputs
Because the arrow keys are not available for moving the cursor when
inputs are opaque, we want to place the cursor in a uniform, most
generically useful position. Assuming that most input devices offer
"Backspace" to delete the character to the left of the cursor, the most
generic choice should be the very end of the input text.

I was surprised that the 'email' input type is a bit weird when it comes
to cursor movements (other input types are too, but it's obvious that
they are not simple text inputs). It seems that in the HTML spec, it is
assumed that email inputs may also be "structured" inputs, though in
practice they currently are not:

> Currently, all browsers which implement this element implement it as a
standard text input field with basic validation features. The
specification does, however, allow browsers latitude on this.

https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input/email