Add basic tests for opaque vs normal inputs
At least in Firefox Cypress does not seem to be able to faithfully model
the behaviour of the text cursor in inputs, the cursor seemed to just
stay at the beginning of the input text even when an arrow key press
should have moved it within the text, and did so in manual testing
across browsers.

The test is therefore limited to establishing that opaque mode triggers
spatial navigation immediately on input fields, whereas normal mode does
not.