Use checkVisibility when filtering for focusables
`Element.checkVisibility()` is now available in latest versions of all
major browsers. Where it is not available, we can at least roughly
approximate it by recursing up the DOM tree and checking for close
analogues of the evaluated style properties.

This also factors in the recent `content-visibility` CSS property where
supported and used.